+++ src/internal/quo.h
@@ -10,6 +10,10 @@ sexp* rlang_get_expression(sexp* x, sexp* alternate);
 sexp* rlang_quo_get_env(sexp* quo);
 sexp* rlang_quo_get_expr(sexp* quo);
 
+static inline sexp* rlang_quo_get_expr_(sexp* quo) {
+  return r_node_cadr(quo);
+}
+
 void check_quosure(sexp* x);
 bool quo_is_missing(sexp* quo);
 bool quo_is_symbol(sexp* quo);